lab7/pyproject.toml
[project]
name = "MLOps_course_AGH_lab_07"
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "ipykernel>=7.1.0",
    "numpy",
    "onnx",
    "onnxruntime",
    "onnxscript>=0.5.6",
    "sentence-transformers>=5.1.2",
    "torch",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu121", marker = "sys_platform == 'linux' or sys_platform == 'win32'" }
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
